+++ tests/test_managers.py
@@ -94,9 +94,7 @@ class TestResourceManager(unittest.TestCase):
 
     def test_create_empty_resource(self):
         project = self.redmine.project.new()
-        defaults = dict((include, None) for include in project._includes)
-        defaults.update(dict((relation, None) for relation in project._relations))
-        self.assertEqual(project._attributes, defaults)
+        self.assertEqual(repr(project), '<redmine.resources.Project #0 "">')
 
     @mock.patch('redmine.requests.put')
     def test_update_resource(self, mock_put):